feat(models): restructure OpenAI client models and add codecov configuration
- Moved OpenAI client models into a dedicated package.
- Implemented new data classes for ChatRequest, ChatResponse, Message, and others.
- Updated imports across all relevant files to reflect the new structure.
- Added a codecov.yml configuration to enable coverage reporting while ignoring ChatCompletion.kt.
