+++ codecov.yml
@@ -1,0 +1,4 @@
+codecov:
+  enabled: true
+  ignore:
+    - "src/main/kotlin/openai/clients/data/ChatCompletion.kt"
