+++ src/main/kotlin/openai/clients/OpenAiClientImpl.kt
@@ -10,6 +10,9 @@ import kotlinx.serialization.Serializable
 import no.uyqn.config.Configuration
 import no.uyqn.config.EnvironmentalVariable
 import no.uyqn.openai.OpenAiModel
+import no.uyqn.openai.clients.models.ChatRequest
+import no.uyqn.openai.clients.models.ChatResponse
+import no.uyqn.openai.clients.models.Message
 
 class OpenAiClientImpl(
     private val configuration: Configuration,
